# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bijou AI WhatsApp Enterprise",
    "version": "2.2.0",
    "description": "Enterprise-grade AI customer support for WhatsApp with multi-language support, human escalation, and Google Sheets integration. Built for property agents and SMEs.",
    "contact": {
      "name": "Bijou AI Support Team",
      "email": "support@mybijou.xyz",
      "url": "https://app.mybijou.xyz/static/help.html"
    },
    "license": {
      "name": "Commercial License",
      "url": "https://mybijou.xyz/terms"
    }
  },
  "paths": {
...
    "/onboard/{token}": {
      "get": {
        "summary": "Serve Onboarding",
        "description": "Serve onboarding page for property agent signup",
        "operationId": "serve_onboarding_onboard__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/onboarding/signup": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Signup Property Agent",
        "description": "Step 1: Create new tenant and initiate WhatsApp session\n\nReturns onboarding URL with token for QR display page",
        "operationId": "signup_property_agent_api_onboarding_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/onboarding/status/{token}": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Get Onboarding Status",
        "description": "Step 2: Check WhatsApp connection status\n\nCalled by onboarding page every 3 seconds to detect when QR is scanned",
        "operationId": "get_onboarding_status_api_onboarding_status__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/onboarding/qr/{token}": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Get Qr Code",
        "description": "Step 3: Proxy QR code image from WhatsApp bridge\n\nReturns PNG image that updates when tenant_id changes",
        "operationId": "get_qr_code_api_onboarding_qr__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/onboarding/complete/{token}": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Complete Onboarding",
        "description": "Step 4: Mark onboarding as complete\n\nCalled when WhatsApp connection is confirmed",
        "operationId": "complete_onboarding_api_onboarding_complete__token__post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SignupRequest": {
        "properties": {
          "business_name": {
            "type": "string",
            "title": "Business Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "business_name",
          "email",
          "phone"
        ],
        "title": "SignupRequest",
        "description": "Property agent signup form data"
      },
      "SignupResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "onboarding_url": {
            "type": "string",
            "title": "Onboarding Url"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tenant_id",
          "onboarding_url",
          "message"
        ],
        "title": "SignupResponse",
        "description": "Response after successful signup"
      },
      "StatusResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "business_name": {
            "type": "string",
            "title": "Business Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "whatsapp_connected": {
            "type": "boolean",
            "title": "Whatsapp Connected"
          },
          "onboarding_completed": {
            "type": "boolean",
            "title": "Onboarding Completed"
          },
          "whatsapp_jid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Jid"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "business_name",
          "status",
          "whatsapp_connected",
          "onboarding_completed",
          "created_at"
        ],
        "title": "StatusResponse",
        "description": "WhatsApp connection status"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}